Some minor fixes

+++ bukkit0/src/main/java/com/boydti/fawe/bukkit/v0/BukkitQueue_All.java
@@ -38,7 +38,7 @@ public class BukkitQueue_All extends BukkitQueue_0<Chunk, Chunk, Chunk> {
     }
 
     @Override
-    public void setHeightMap(FaweChunk chunk, int[] heightMap) {
+    public void setHeightMap(FaweChunk chunk, byte[] heightMap) {
         // Do nothing
     }
